Remove diem types
