+++ schemars/src/json_schema_impls/mod.rs
@@ -65,5 +65,4 @@ mod url;
 #[cfg(feature = "uuid")]
 mod uuid;
 mod wrapper;
-
-mod diem_types;
+mod multi_addr;
